Build the client modules into separate files so as to make it possible to avoid circular dependencies
