+++ runtime/src/app/goto.ts
@@ -1,4 +1,4 @@
-import { cid, history, navigate, select_target } from '../router';
+import { cid, history, navigate, select_target } from './router';
 
 export default function goto(
 		href: string,
